requirements.txt
torchrl
hydra-core
tqdm
wandb
torchvision

# dm_control
torchrl[dm_control]
torchrl[gymnasium]

gymnasium
dm_control


#swig
gymnasium[box2d]

moviepy